python-code/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
pymongo==3.12.0
numpy==1.21.2
scikit-learn==0.24.2
langdetect==1.0.9
tensorflow==2.6.0
langchain-groq==0.0.1
opencv-python==4.5.3.56
Pillow==8.3.2
SpeechRecognition==3.8.1
pyttsx3==2.90
pytesseract==0.3.8
python-multipart
werkzeug
deepface==0.0.75
requests==2.26.0
langchain==0.1.0
langchain-community==0.0.1
transformers==4.11.3
gunicorn==20.1.0

